+++ bot/bots/scenes/books/search.js
@@ -1,8 +1,8 @@
 const BaseScene = require('telegraf/scenes/base');
 
-const MAX_RETRIES = 2;
+const MAX_RETRIES = 5;
 
-async function textSearch(ctx, params, retry = 1) {
+async function textSearch(ctx, params, retry = 1, message = false) {
     const {getBookList, sceneCode, discoverSceneCode} = params;
 
     let query = ctx.update.message && ctx.update.message.text
@@ -13,9 +13,11 @@ async function textSearch(ctx, params, retry = 1) {
         return ctx.scene.reenter();
     }
 
-    ctx.perfStart('searchMsg');
-    let message = await ctx.reply('Ищу...');
-    ctx.perfStop('searchMsg');
+    if (retry === 1) {
+        ctx.perfStart('searchMsg');
+        message = await ctx.reply('Ищу...');
+        ctx.perfStop('searchMsg');
+    }
 
     ctx.perfStart('queryBookList');
     let items = await getBookList(query);
@@ -23,16 +25,18 @@ async function textSearch(ctx, params, retry = 1) {
 
     if (items === false) {
         if (retry < MAX_RETRIES) {
-            return textSearch(ctx, params, ++retry);
+            return textSearch(ctx, params, ++retry, message);
         }
         else {
             return ctx.reply('Ошибка при поиске. Попробуйте еще раз');
         }
     }
 
-    ctx.perfStart('deleteSearchMsg');
-    await ctx.deleteMessage(message.message_id);
-    ctx.perfStop('deleteSearchMsg');
+    if (message && message.message_id) {
+        ctx.perfStart('deleteSearchMsg');
+        await ctx.deleteMessage(message.message_id);
+        ctx.perfStop('deleteSearchMsg');
+    }
 
     await ctx.perfCommit();
 
@@ -50,13 +54,23 @@ module.exports = function (params) {
     const scene = new BaseScene(sceneCode);
 
     scene.enter(async ctx => {
+        let newMessage = ctx.scene.state.newMessage || false;
         ctx.session.delaySubscribeCheck = false;
         ctx.perfStart('searchIntro');
-        await ctx.safeReply(
-            ctx => ctx.editMessageText('Напишите запрос и я поищу подходящие книги'),
-            ctx => ctx.reply('Напишите запрос и я поищу подходящие книги'),
-            ctx
-        );
+        if (newMessage) {
+            await ctx.safeReply(
+                ctx => ctx.reply('Напишите запрос и я поищу подходящие книги'),
+                false,
+                ctx
+            );
+        }
+        else {
+            await ctx.safeReply(
+                ctx => ctx.editMessageText('Напишите запрос и я поищу подходящие книги'),
+                ctx => ctx.reply('Напишите запрос и я поищу подходящие книги'),
+                ctx
+            );
+        }
         ctx.perfStop('searchIntro');
         await ctx.perfCommit();
     });
